requirements.txt
# Core ML Libraries
lightgbm==4.1.0
catboost==1.2.2
pandas==2.1.4
polars==0.20.2
numpy==1.26.2
scikit-learn==1.4.0

# MLflow for experiment tracking
mlflow==2.11.3
mlflow.sklearn

# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Data processing
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# AWS SDK (for deployment)
boto3==1.34.28

# Azure SDK (for deployment)
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

